requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
eventlet==0.31.1
greenlet==1.1.1
Pillow==8.3.1
python-socketio==5.4.0
psycopg2-binary
dj-database-url
django-cors-headers
